Add library for communication with the motor controller
This commit also adds a "fake" implementation of the required UART
abstraction needed by the third party library. That is, an
implementation that has the correct API for building with, but that
won't be useful at runtime.

Additionally, this commit adds support for making `clang-format` ignore
specified directories, and then proceeds to make it ignore the
third-party vesc-uart library.

This commit also updates `clang-tidy` so that system include are
correctly ordered when using `--spawn_strategy=local`.

Change-Id: I350bf62912454d7b815a9dedcc048a63c2107c46