Add serial port example
Add an example connecting to a serial port and writing a message. This
is tested by creating a virtual serial port with the output written to
file.

Bazel test interaction with the filesystem prevents creation of a RW
device. As a result, `//test:virtual_serial_port_test` is defined as a
`sh_binary` that can be run with:

  bazel run //test:virtual_serial_port_test

Change-Id: I1be9cb2e8b13fa972084d66fa2c6ddffb23ae3cc