Add example to build and link against `libi2c` (#15)
Define `//examples:i2c` which intended to test building a program that
links against `libi2c`. This example is marked incompatible if `libi2c`
is not detected on the host platform during workspace configuration.

Linux x86-64 is configured to use hermetic toolchains with a hermetic
sysroot. As a result, system libraries need to be explicitly specified
as `/usr/include` and `/usr/lib` are not searched for headers and
libraries. When building for this host, `libi2c` is explicitly defined
to allow use.

Change-Id: I232a877e7e7bb6afbbc5a956e5eaba75488e190a