Add reaction wheel balance application
This commit adds and application in the state that it was when the
hackathon at TU Delft bike lab was completed.

In addition to adding the application, this commit also revert the MEMS
parameters to those of the demo application. Additionally, the sensor
position was moved before the very last runs and thus the evaluation of
the roll angle from the accelerometer is updated to reflect that.