Add reaction wheel balance application (#23)
This commit adds an application in the state that it was when the
hackathon at TU Delft bike lab was completed.

In addition to adding the application, this commit also reverts the MEMS
parameters to those of the demo application. Additionally, the sensor
position was moved before the very last runs and thus the evaluation of
the roll angle from the accelerometer is updated to reflect that.

Co-authored-by: Björn Andersson <andebjor@gmail.com>